refactor : renamed files
dataContext => productContext
filterContext => dataContext
handleFilters => reducer
