+++ src/reducer/index.js
@@ -1,0 +1,1 @@
+export { reducer } from "./reducer";
